requirements.txt
numpy
numpy-stl
pandas
pyvista
rasterio
scipy
trimesh
vtk
